Time: 699 ms (10.35%), Space: 254.5 MB (7.43%) - LeetHub

+++ 204-count-primes/204-count-primes.cpp
@@ -4,7 +4,7 @@ public:
         vector<long> prime(n, 1);
         for(int i=2; i*i<n; i++){
             if(prime[i]){
-                for(long j=i*i; j<n; j+=i){
+                for(int j=i*i; j<n; j+=i){
                     prime[j]=0;
                 }
             }
